Move `ObservationToken` and `ProductImagesProvider` to their own files.
