+++ WooCommerce/Classes/Tools/ObservationToken.swift
@@ -1,0 +1,13 @@
+/// An observation token that contains a closure when the observation is cancelled.
+///
+final class ObservationToken {
+    private let onCancel: () -> Void
+
+    init(onCancel: @escaping () -> Void) {
+        self.onCancel = onCancel
+    }
+
+    func cancel() {
+        onCancel()
+    }
+}
